app/requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==2015.04.28
cffi==1.1.0
cryptography==0.9
enum34==1.0.4
idna==2.0
ipaddress==1.0.7
ipython==3.1.0
itsdangerous==0.24
ndg-httpsclient==0.4.0
pushbullet.py==0.8.1
pyOpenSSL==0.15.1
pyasn1==0.1.7
pycparser==2.13
python-magic==0.4.6
requests==2.7.0
six==1.9.0
websocket-client==0.31.0
wsgiref==0.1.2
